app/pyproject.toml
# Defaults
[tool.poetry]
name = "karaoke-app"
version = "0.1.0"
description = ""
authors = ["Andrews <andyarok@gmail.com>"]
license = "MIT"
#packages=["app"]

# Dependencies
[tool.poetry.dependencies]
python = "^3.10"
# google sdk for pub/sub
google-cloud-pubsub = "^2.7.0"
# google sdk for storage
google-cloud-storage = "^1.42.0"
fastapi = "^0.111.0"
firebase-admin = "^6.5.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
